This is The configuration for the most exciting tiling WM - Qtile
Qtile is a full-featured, hackable tiling window manager written in Python. Qtile is simple, small, and extensible. It's easy to write your own layouts, widgets, and built-in commands. It is written and configured entirely in Python, which means you can leverage the full power and flexibility of the language to make it fit your needs.

If you have'nt installed Qtile Yet refer to this link:
http://docs.qtile.org/en/stable/manual/install/index.html

All of the configuration is in config.py please install the libraries via pip3[Python package installer]
You can get pip here:
https://pypi.org/project/pip/

All of the autostart configuration is in autostart.sh, please customize it accordingly!!!